feat(AIF): implemented AIF auto detection tools.
Required some major architecture reformatting, including splitting fitting out of mdm_DCEVoxel (solving issue #24 in the process).

In addition to the new AIF fitting tool, an AIF ROI map can be supplied as input to madym_DCE, from which an AIF will be computed on the fly using the average C(t) of voxels specified in the map.

Added AIF and ROI saving to the output of madym_DCE.